requirement.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
tensorflow==2.16.1
pillow==10.1.0
numpy==1.26.1
pip install fastapi==0.104.1 uvicorn==0.24.0 python-multipart==0.0.6 tensorflow==2.16.1 pillow==10.1.0 numpy==1.26.1